requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
Flask-WTF==1.2.1
Flask-Session==0.5.0